#15 refactor: whitespace-nowrap css 추가, 버튼 컴포넌트 이동
- 화면을 줄였을 때 글자들이 영역 밖을 벗어나지 않게 하기 위해 whitespace-nowrap css 추가
- 큰 서치바의  버튼을 재사용 관점에서 봤을 때 재사용 가능성이 없어 버튼 컴포넌트를 서치바 컴포넌트로 이동
