+++ src/App.jsx
@@ -4,13 +4,15 @@ import Container from './components/Container/Container.jsx'
 import RoomList from './components/RoomList/RoomList.jsx'
 import SearchBar from './components/SearchBar/SearchBar'
 import Footer from './components/Footer/Footer.jsx'
+import SmallSearchBar from './components/SearchBar/SmallSearchBar.jsx'
 function App() {
   return (
     <>
-      <Container />
+      {/* <Container /> */}
       <SearchBar />
-      <RoomList></RoomList>
-      <Footer />
+      {/* <SmallSearchBar></SmallSearchBar> */}
+      {/* <RoomList></RoomList>
+      <Footer /> */}
     </>
   )
 }
