--- src/components/SearchBar/SearchButton.jsx
@@ -1,11 +1,0 @@
-const SearchButtonArea = () => (
-    <div className="Group2 w-1/6 h-12 relative py-1.5 ">
-      <svg width="48" height="49" viewBox="0 0 48 49" fill="none" xmlns="http://www.w3.org/2000/svg">
-        <circle  cx="24" cy="24.5" r="24" fill="#FF385C"/>
-        <circle cx="22.5" cy="23" r="5.98611" stroke="white" />
-        <line x1="31.9492" y1="33.0508" x2="25.5852" y2="26.6869" stroke="white" />
-      </svg>
-    </div>
-  );
-
-  export default SearchButtonArea;
